URLPattern: Add wildcard '*' support.
This CL adds support for a wildcard `*` character in patterns.  It is
treated as being equivalent to the `(.*)` unnamed regexp group.  This
means it can have its own following modifiers like `*?` which can affect
how prefix characters are treated.

This CL deviates from the current upstream path-to-regexp v6.2.0, but
we plan to upstream this change.  Initial discussions suggest that its
possible this will be accepted.

Bug: 1141510
Change-Id: I386b03e8bb05aab9586464986faf002e824e189e